djangoweb/requirements.txt
Django>=5.2,<5.3
psycopg2-binary>=2.9.10,<2.10
pillow>=11.2.1,<11.3
python-dotenv>=1.1.0,<1.2